+++ src/Oro/Bundle/AssetBundle/Command/OroAssetsBuildCommand.php
@@ -266,9 +266,6 @@ HELP
             $command[] = self::BUILD_DIR.'/node_modules/webpack/bin/webpack.js';
         }
 
-        if ($input->getArgument('theme')) {
-            $command[] = '--env theme='.$input->getArgument('theme');
-        }
         if (true === $input->getOption('no-debug') || 'prod' === $input->getOption('env')) {
             $command[] = '--mode=production';
         }
@@ -282,34 +279,32 @@ HELP
         } else {
             $verbosity = self::WEBPACK_VERBOSITY_MAP[$output->getVerbosity()];
         }
-        if ($verbosity) {
-            $command[] = '--env';
-            $command[] = 'stats=' . $verbosity;
-        }
+
         $command[] = '--env';
         $command[] = 'symfony='.$input->getOption('env');
 
+        if ($verbosity) {
+            $command[] = 'stats=' . $verbosity;
+        }
+        if ($input->getArgument('theme')) {
+            $command[] = 'theme='.$input->getArgument('theme');
+        }
         if ($input->getOption('skip-css')) {
-            $command[] = '--env';
             $command[] = 'skipCSS';
         }
         if ($input->getOption('skip-js')) {
-            $command[] = '--env';
             $command[] = 'skipJS';
         }
         if (!$input->getOption('with-babel')) {
-            $command[] = '--env';
             $command[] = 'skipBabel';
         }
         if ($input->getOption('skip-sourcemap')) {
-            $command[] = '--env';
             $command[] = 'skipSourcemap';
         }
         if ($input->getOption('skip-rtl')) {
-            $command[] = '--env.skipRTL';
+            $command[] = 'skipRTL';
         }
         if ($input->getOption('analyze')) {
-            $command[] = '--env';
             $command[] = 'analyze';
         }